done with set reporting and leaderboard

+++ smash_ladder/apps.py
@@ -5,13 +5,13 @@ from django.apps import AppConfig
 
 @dataclass(frozen=True)
 class RatingServiceConfig:
-    default_rating: int
-    k_factor: int
+    default_rating: float
+    k_factor: float
 
 
 class SmashLadderConfig(AppConfig):
     name = 'smash_ladder'
     rating_config = RatingServiceConfig(
-        default_rating=1200,
-        k_factor=30
+        default_rating=1200.0,
+        k_factor=30.0
     )
